# java/helidon/src/main/java/org/openapitools/server/api/UserService.java
/*
 * Swagger Petstore - OpenAPI 3.1
 * This is a sample Pet Store Server based on the OpenAPI 3.1 specification.  You can find out more about Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach! You can now help us improve the API whether it's by making changes to the definition itself or to the code. That way, with time, we can improve the API in general, and expose some of the new features in OAS3.  Some useful links: - [The Pet Store repository](https://github.com/swagger-api/swagger-petstore) - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
 *
 * The version of the OpenAPI document: 1.0.12
 * Contact: apiteam@swagger.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

package org.openapitools.server.api;

import org.openapitools.server.model.Error;
import java.util.List;
import java.time.OffsetDateTime;
import java.util.Optional;
import org.openapitools.server.model.User;

import jakarta.ws.rs.*;

import java.io.InputStream;
import java.util.Map;
import java.util.List;
import jakarta.validation.constraints.*;
import jakarta.validation.Valid;

@Path("/user")
@jakarta.annotation.Generated(value = "org.openapitools.codegen.languages.JavaHelidonServerCodegen", date = "2026-06-09T23:01:38.088546400-04:00[America/Toronto]", comments = "Generator version: 7.23.0")
public interface UserService {

    @POST
    @Consumes({ "application/json", "application/xml", "application/x-www-form-urlencoded" })
    @Produces({ "application/json", "application/xml" })
    User createUser(@Valid User user);

    @POST
    @Path("/createWithList")
    @Consumes({ "application/json" })
    @Produces({ "application/json", "application/xml" })
    User createUsersWithListInput(@Valid List<@Valid User> user);

    @DELETE
    @Path("/{username}")
    @Produces({ "application/json" })
    void deleteUser(@PathParam("username") String username);

    @GET
    @Path("/{username}")
    @Produces({ "application/json", "application/xml" })
    User getUserByName(@PathParam("username") String username);

    @GET
    @Path("/login")
    @Produces({ "application/xml", "application/json" })
    String loginUser(@QueryParam("username") String username, @QueryParam("password") String password);

    @GET
    @Path("/logout")
    void logoutUser();

    @PUT
    @Path("/{username}")
    @Consumes({ "application/json", "application/xml", "application/x-www-form-urlencoded" })
    @Produces({ "application/json" })
    void updateUser(@PathParam("username") String username, @Valid User user);
}